Freefall files in fortran
Coding in fortran makes the simulations faster. Without parallelization, the code would still have been faster than the one used in python. This exercise is a great demonstration of fortran's capacity to process code. Credits to CAst at the LMU for the course and providing the skeleton on which the code was devloped and altered.